+++ examples/valid/valid.hs
@@ -5,10 +5,16 @@
 -- Copyright   : Copyright (c) 2019-2023 Travis Cardwell
 -- License     : MIT
 --
--- 'TTC.valid' is used to create validated constants.  The sample username
--- is validated at compile-time.
+-- 'TTC.valid' is used to create validated constants.  The type must have a
+-- `Lift` instance since this is implemented using typed Template Haskell.
+-- When using the `OverloadedStrings` extension, you do not even have to write
+-- the function name, at the expense of documentation.  With GHC 9, you can
+-- even leave off the parenthesis.  The sample usernames are validated at
+-- compile-time.
 ------------------------------------------------------------------------------
 
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
 
 module Main (main) where
@@ -21,10 +27,22 @@ import Username (Username)
 
 ------------------------------------------------------------------------------
 
-sample :: Username
-sample = $$(TTC.valid "tcard")
+sample1 :: Username
+-- This syntax works with all supported versions of GHC.
+sample1 = $$(TTC.valid "tcard")
+
+sample2 :: Username
+#if __GLASGOW_HASKELL__ >= 900
+-- This syntax only works with GHC 9.
+sample2 = $$"alice"
+#else
+-- This syntax works with all supported versions of GHC.
+sample2 = $$("alice")
+#endif
 
 ------------------------------------------------------------------------------
 
 main :: IO ()
-main = print sample
+main = do
+    print sample1
+    print sample2